chore: [AICP-43] rename the file and move slide image to the molecules
