+++ src/assets/images/SlideImages/index.ts
