+++ src/components/molecules/SlideImages/SlideImages.scss
