+++ src/components/molecules/index.ts
@@ -3,3 +3,4 @@ export * from './DropProfile/DropProfile';
 export * from './Modal/OpenModal';
 export * from './Modal/Modal';
 export * from './EmptyData/EmptyData';
+export * from './SlideImages/SlideImages';
